+++ modules/user_kustomizations/main.tf
@@ -14,13 +14,14 @@ module "user_kustomization_set" {
 
   pre_commands_string  = var.kustomizations_map[each.key].pre_commands
   post_commands_string = var.kustomizations_map[each.key].post_commands
+  replacement_triggers = var.replacement_triggers
 }
 
 resource "terraform_data" "kustomization_user_deploy" {
 
-  triggers_replace = {
+  triggers_replace = merge({
     kustomization_shas = nonsensitive(sha256(yamlencode(module.user_kustomization_set)))
-  }
+  }, var.replacement_triggers)
 
   connection {
     user           = var.ssh_connection.user
@@ -41,11 +42,6 @@ resource "terraform_data" "kustomization_user_deploy" {
       #!/bin/bash
       set -e
 
-      function cleanup {
-        echo "Cleaning up ${local.base_destination_folder}..."
-        rm -rf ${local.base_destination_folder}
-      }
-      trap cleanup EXIT
       sorted_dest_keys=$(printf '%s\n' ${join(" ", local.destination_keys)} | sort -n | tr '\n' ' ')
 
       for dest_key in $sorted_dest_keys; do